requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas==1.2.2
pandas-profiling==2.11.0
scikit-learn==0.22
scikit-image==0.18.1

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs==2021.6.1
google-cloud-storage==1.39.0
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor

# docker
fastapi
uvicorn
